Rename tdx-profile to tdx
