+++ coev/tdx/testcases/ce-pce-evidence.cbor
