+++ coev/tdx/testcases/ce-qe-evidence.cbor
