+++ coev/tdx/testcases/ce-seam-evidence.cbor
