+++ coev/tdx/testcases/regen-from-src.sh
